backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
django-channels[daphne]>=4.0
channels-redis>=4.1
mysqlclient>=2.2
celery>=5.3
redis>=5.0
Pillow>=10.0
numpy>=1.26
scipy>=1.11
gunicorn>=21.2
